distributed-os-lab1/requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.61.0
grpcio==1.76.0
grpcio-tools==1.76.0
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
